unify the two ways of thinking
